+++ lib/src/truecaller.dart
@@ -38,8 +38,11 @@ import 'scope_options.dart';
 import 'truecaller_callback.dart';
 
 class TruecallerSdk {
-  static const MethodChannel _methodChannel = const MethodChannel('tc_method_channel');
-  static const EventChannel _eventChannel = const EventChannel('tc_event_channel');
+  static const MethodChannel _methodChannel =
+      const MethodChannel('tc_method_channel');
+  static const EventChannel _eventChannel =
+      const EventChannel('tc_event_channel');
+  static Stream<TruecallerSdkCallback> _callbackStream;
 
   /// This method has to be called before anything else. It initializes the SDK with the
   /// customizable options which are all optional and have default values as set below in the method
@@ -48,9 +51,10 @@ class TruecallerSdk {
   /// 1. [TruecallerSdkScope.SDK_OPTION_WITHOUT_OTP] i.e only Truecaller users
   /// 2. [TruecallerSdkScope.SDK_OPTION_WITH_OTP] i.e both Truecaller and Non-truecaller users
   ///
-  /// NOTE: As of truecaller_sdk 0.0.1, only
+  /// NOTE: In truecaller_sdk 0.0.1, only
   /// [TruecallerSdkScope.SDK_OPTION_WITHOUT_OTP] is supported
-  ///
+  /// In truecaller_sdk 0.0.2 and onwards, both [TruecallerSdkScope.SDK_OPTION_WITHOUT_OTP] and
+  /// [TruecallerSdkScope.SDK_OPTION_WITH_OTP] are supported
   /// [consentMode] determines which kind of consent screen you want to show to the user.
   /// [consentTitleOptions] is applicable only for [TruecallerSdkScope.CONSENT_MODE_POPUP]
   /// and [TruecallerSdkScope.CONSENT_MODE_FULLSCREEN] and it sets the title prefix
@@ -69,10 +73,13 @@ class TruecallerSdk {
   static initializeSDK(
           {@required int sdkOptions,
           int consentMode: TruecallerSdkScope.CONSENT_MODE_BOTTOMSHEET,
-          int consentTitleOptions: TruecallerSdkScope.SDK_CONSENT_TITLE_GET_STARTED,
+          int consentTitleOptions:
+              TruecallerSdkScope.SDK_CONSENT_TITLE_GET_STARTED,
           int footerType: TruecallerSdkScope.FOOTER_TYPE_SKIP,
-          int loginTextPrefix: TruecallerSdkScope.LOGIN_TEXT_PREFIX_TO_GET_STARTED,
-          int loginTextSuffix: TruecallerSdkScope.LOGIN_TEXT_SUFFIX_PLEASE_LOGIN,
+          int loginTextPrefix:
+              TruecallerSdkScope.LOGIN_TEXT_PREFIX_TO_GET_STARTED,
+          int loginTextSuffix:
+              TruecallerSdkScope.LOGIN_TEXT_SUFFIX_PLEASE_LOGIN,
           int ctaTextPrefix: TruecallerSdkScope.CTA_TEXT_PREFIX_USE,
           String privacyPolicyUrl: "",
           String termsOfServiceUrl: "",
@@ -99,11 +106,14 @@ class TruecallerSdk {
   /// [TruecallerSdkScope.SDK_OPTION_WITHOUT_OTP], you can check if the Truecaller app is
   /// present on the user's device or whether the user has a valid account state or not by using
   /// the following method
-  static Future<bool> get isUsable async => _methodChannel.invokeMethod('isUsable');
+  static Future<bool> get isUsable async =>
+      _methodChannel.invokeMethod('isUsable');
 
   /// After checking [isUsable], you can show the Truecaller profile verification dialog
   /// anywhere in your app flow by calling the following method
-  static get getProfile async => await _methodChannel.invokeMethod('getProfile');
+  /// The result will be returned asynchronously via [streamCallbackData] stream
+  static get getProfile async =>
+      await _methodChannel.invokeMethod('getProfile');
 
   /// Once you call [getProfile], you can listen to this stream to determine the result of the
   /// action taken by the user.
@@ -114,41 +124,122 @@ class TruecallerSdk {
   /// [TruecallerSdkCallbackResult.verification] will be returned only when using
   /// [TruecallerSdkScope.SDK_OPTION_WITH_OTP] which indicates to verify the user
   /// manually, so this is not applicable for truecaller_sdk 0.0.1
-  static Stream<TruecallerSdkCallback> get streamCallbackData =>
-      _eventChannel.receiveBroadcastStream().map((event) {
+  static Stream<TruecallerSdkCallback> get streamCallbackData {
+    if (_callbackStream == null) {
+      _callbackStream = _eventChannel
+          .receiveBroadcastStream()
+          .map<TruecallerSdkCallback>((value) {
         TruecallerSdkCallback callback = new TruecallerSdkCallback();
-        var resultHashMap = HashMap<String, String>.from(event);
+        var resultHashMap = HashMap<String, String>.from(value);
         switch (resultHashMap["result"].enumValue()) {
           case TruecallerSdkCallbackResult.success:
             callback.result = TruecallerSdkCallbackResult.success;
-            Map profileMap = jsonDecode(resultHashMap["data"]);
-            TruecallerUserProfile truecallerUserProfile =
-                TruecallerUserProfile.fromJson(profileMap);
-            callback.profile = truecallerUserProfile;
+            _insertProfile(callback, resultHashMap["data"]);
             break;
           case TruecallerSdkCallbackResult.failure:
             callback.result = TruecallerSdkCallbackResult.failure;
-            Map errorMap = jsonDecode(resultHashMap["data"]);
-            TruecallerError truecallerError = TruecallerError.fromJson(errorMap);
-            callback.error = truecallerError;
+            _insertError(callback, resultHashMap["data"]);
             break;
           case TruecallerSdkCallbackResult.verification:
             callback.result = TruecallerSdkCallbackResult.verification;
+            _insertError(callback, resultHashMap["data"]);
+            break;
+          case TruecallerSdkCallbackResult.missedCallInitiated:
+            callback.result = TruecallerSdkCallbackResult.missedCallInitiated;
+            callback.ttl = resultHashMap["data"];
+            break;
+          case TruecallerSdkCallbackResult.missedCallReceived:
+            callback.result = TruecallerSdkCallbackResult.missedCallReceived;
+            break;
+          case TruecallerSdkCallbackResult.otpInitiated:
+            callback.result = TruecallerSdkCallbackResult.otpInitiated;
+            callback.ttl = resultHashMap["data"];
+            break;
+          case TruecallerSdkCallbackResult.otpReceived:
+            callback.result = TruecallerSdkCallbackResult.otpReceived;
+            callback.otp = resultHashMap["data"];
+            break;
+          case TruecallerSdkCallbackResult.verifiedBefore:
+            callback.result = TruecallerSdkCallbackResult.verifiedBefore;
+            _insertProfile(callback, resultHashMap["data"]);
+            break;
+          case TruecallerSdkCallbackResult.verificationComplete:
+            callback.result = TruecallerSdkCallbackResult.verificationComplete;
+            callback.accessToken = resultHashMap["data"];
+            break;
+          case TruecallerSdkCallbackResult.exception:
+            callback.result = TruecallerSdkCallbackResult.exception;
+            Map exceptionMap = jsonDecode(resultHashMap["data"]);
+            TruecallerException exception =
+                TruecallerException.fromJson(exceptionMap);
+            callback.exception = exception;
             break;
           default:
-            throw ArgumentError('${resultHashMap["result"]} is not a valid result');
+            throw PlatformException(
+                code: "1010",
+                message: "${resultHashMap["result"]} is not a valid result");
         }
         return callback;
       });
+    }
+    return _callbackStream;
+  }
+
+  static _insertProfile(TruecallerSdkCallback callback, String data) {
+    Map profileMap = jsonDecode(data);
+    TruecallerUserProfile profile = TruecallerUserProfile.fromJson(profileMap);
+    callback.profile = profile;
+  }
+
+  static _insertError(TruecallerSdkCallback callback, String data) {
+    // onVerificationRequired has nullable error, hence null check
+    if (data != "null") {
+      Map errorMap = jsonDecode(data);
+      TruecallerError truecallerError = TruecallerError.fromJson(errorMap);
+      callback.error = truecallerError;
+    }
+  }
 
   /// To customise the look and feel of the verification consent screen as per your app theme, add
   /// the following lines before calling the [getProfile] method.
   /// NOTE: It's not applicable for [TruecallerSdkScope.CONSENT_MODE_BOTTOMSHEET]
-  static get setDarkTheme async => await _methodChannel.invokeMethod('setDarkTheme');
+  static get setDarkTheme async =>
+      await _methodChannel.invokeMethod('setDarkTheme');
 
   /// To customise the profile dialog in any of the supported Indian languages, add the
   /// following lines before calling the [getProfile] method with [locale] of your choice.
   /// NOTE: Default value is en
   static setLocale(String locale) async =>
       await _methodChannel.invokeMethod('setLocale', {"locale": locale});
+
+  /// This method will initiate manual verification of [phoneNumber] asynchronously for Indian
+  /// numbers only as of truecaller_sdk 0.0.2 so that's why default countryISO is set to "IN".
+  /// The result will be returned asynchronously via [streamCallbackData] stream
+  /// Check [TruecallerSdkCallbackResult] to understand the different verifications states.
+  /// This method may lead to verification with a SMS Code (OTP) or verification with a CALL,
+  /// or if the user is already verified on the device, will get the call back as
+  /// [TruecallerSdkCallbackResult.verifiedBefore] in [streamCallbackData]
+  static requestVerification(
+          {@required String phoneNumber, String countryISO: "IN"}) async =>
+      await _methodChannel.invokeMethod(
+          'requestVerification', {"ph": phoneNumber, "ci": countryISO});
+
+  /// Call this method after [requestVerification] to complete the verification if the number has
+  /// to be verified with a missed call.
+  /// i.e call this method only when you receive [TruecallerSdkCallbackResult.missedCallReceived]
+  /// in [streamCallbackData].
+  /// To complete verification, it is mandatory to pass [firstName] and [lastName] of the user
+  static verifyMissedCall(String firstName, String lastName) async =>
+      await _methodChannel.invokeMethod(
+          'verifyMissedCall', {"fname": firstName, "lname": lastName});
+
+  /// Call this method after [requestVerification] to complete the verification if the number has
+  /// to be verified with an OTP.
+  /// i.e call this method when you receive either [TruecallerSdkCallbackResult.otpInitiated] or
+  /// [TruecallerSdkCallbackResult.otpReceived] in [streamCallbackData].
+  /// To complete verification, it is mandatory to pass [firstName] and [lastName] of the user
+  /// with the [otp] code received over SMS
+  static verifyOtp(String firstName, String lastName, String otp) async =>
+      await _methodChannel.invokeMethod(
+          'verifyOtp', {"fname": firstName, "lname": lastName, "otp": otp});
 }